requirements.txt
torch
torchvision
PyYAML
tensorboard
kornia
tqdm
einops
Pillow
opencv-python
scipy
numpy
matplotlib
